feat: update header to 1.21.70

+++ src/mc/deps/minecraft_renderer/game/ShadowContext.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class ShadowContext : int {
+enum class ShadowContext : uchar {
     Unknown             = 0,
     ShadowOnly          = 1,
     PlayerShadowMapOnly = 2,
